+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - #50 - DS.bindAll($scope, prop, resourceName, query)
 - #52 - DS.update(resourceName, id, attrs[, options]) (different from DS.save())
 - #54 - Adding functionality to resources
+- #56 - DS.updateAll(resourceName, attrs, params[, options])
 
 ##### 0.8.1 - 02 May 2014